Added another animation to the logo.
Added 4 more frames for the Sovereign logo to add a glitchy effect. The
source files are in a string array and each of them is printed with a
delay to achieve this. Also reverted back to printing the file by line
instead of character beacuse we aren't doing any advanced coloring right
now.